[TASK] Tweak schema migrations to Flow/Party decoupling
This adjusts schema migrations to that it is actually possible to:

* install Flow without Party
* add Party to a "clean" Flow setup

Releases: master, 3.0
Related: FLOW-5
Original-Commit-Hash: 37d7f607329a1d792fd5c21be132aba77ab5f165

+++ TYPO3.Flow/Migrations/Mysql/Version20110923125535.php
@@ -22,11 +22,15 @@ class Version20110923125535 extends AbstractMigration {
 		$this->addSql("ALTER TABLE typo3_flow3_resource_resource ADD  CONSTRAINT typo3_flow3_resource_resource_ibfk_1 FOREIGN KEY (resourcepointer) REFERENCES typo3_flow3_resource_resourcepointer(hash)");
 		$this->addSql("CREATE INDEX IDX_B4D45B323CB65D1 ON typo3_flow3_resource_resource (resourcepointer)");
 
-		$this->addSql("ALTER TABLE typo3_flow3_security_account DROP FOREIGN KEY typo3_flow3_security_account_ibfk_1");
-		$this->addSql("DROP INDEX IDX_44D0753B38110E12 ON typo3_flow3_security_account");
+		if ($this->isPartyPackageInstalled()) {
+			$this->addSql("ALTER TABLE typo3_flow3_security_account DROP FOREIGN KEY typo3_flow3_security_account_ibfk_1");
+			$this->addSql("DROP INDEX IDX_44D0753B38110E12 ON typo3_flow3_security_account");
+		}
 		$this->addSql("ALTER TABLE typo3_flow3_security_account CHANGE party_abstractparty party VARCHAR(40) DEFAULT NULL");
-		$this->addSql("ALTER TABLE typo3_flow3_security_account ADD CONSTRAINT typo3_flow3_security_account_ibfk_1 FOREIGN KEY (party) REFERENCES typo3_party_domain_model_abstractparty(flow3_persistence_identifier)");
-		$this->addSql("CREATE INDEX IDX_65EFB31C89954EE0 ON typo3_flow3_security_account (party)");
+		if ($this->isPartyPackageInstalled()) {
+			$this->addSql("ALTER TABLE typo3_flow3_security_account ADD CONSTRAINT typo3_flow3_security_account_ibfk_1 FOREIGN KEY (party) REFERENCES typo3_party_domain_model_abstractparty(flow3_persistence_identifier)");
+			$this->addSql("CREATE INDEX IDX_65EFB31C89954EE0 ON typo3_flow3_security_account (party)");
+		}
 	}
 
 	/**
@@ -42,10 +46,22 @@ class Version20110923125535 extends AbstractMigration {
 		$this->addSql("ALTER TABLE typo3_flow3_resource_resource ADD CONSTRAINT typo3_flow3_resource_resource_ibfk_1 FOREIGN KEY (flow3_resource_resourcepointer) REFERENCES typo3_flow3_resource_resourcepointer(hash)");
 		$this->addSql("CREATE INDEX IDX_11FFD19FD0275681 ON typo3_flow3_resource_resource (flow3_resource_resourcepointer)");
 
-		$this->addSql("ALTER TABLE typo3_flow3_security_account DROP FOREIGN KEY typo3_flow3_security_account_ibfk_1");
-		$this->addSql("DROP INDEX IDX_65EFB31C89954EE0 ON typo3_flow3_security_account");
+		if ($this->isPartyPackageInstalled()) {
+			$this->addSql("ALTER TABLE typo3_flow3_security_account DROP FOREIGN KEY typo3_flow3_security_account_ibfk_1");
+			$this->addSql("DROP INDEX IDX_65EFB31C89954EE0 ON typo3_flow3_security_account");
+		}
 		$this->addSql("ALTER TABLE typo3_flow3_security_account CHANGE party party_abstractparty VARCHAR(40) DEFAULT NULL");
-		$this->addSql("ALTER TABLE typo3_flow3_security_account ADD CONSTRAINT typo3_flow3_security_account_ibfk_1 FOREIGN KEY (party_abstractparty) REFERENCES typo3_party_domain_model_abstractparty(flow3_persistence_identifier)");
-		$this->addSql("CREATE INDEX IDX_44D0753B38110E12 ON typo3_flow3_security_account (party_abstractparty)");
+		if ($this->isPartyPackageInstalled()) {
+			$this->addSql("ALTER TABLE typo3_flow3_security_account ADD CONSTRAINT typo3_flow3_security_account_ibfk_1 FOREIGN KEY (party_abstractparty) REFERENCES typo3_party_domain_model_abstractparty(flow3_persistence_identifier)");
+			$this->addSql("CREATE INDEX IDX_44D0753B38110E12 ON typo3_flow3_security_account (party_abstractparty)");
+		}
 	}
+
+	/**
+	 * @return boolean
+	 */
+	protected function isPartyPackageInstalled() {
+		return $this->sm->tablesExist(array('typo3_party_domain_model_abstractparty'));
+	}
+
 }
